FMZ-ENGLISH-TRAINER-003 Add features to summarize irregular verbs and grade statistics
This commit adds functionality which allows to view the current status of learned/unlearned irregular verbs and provides grade statistics for users. Implemented a grade system which includes a DTO and a Mapper. Added a CSS class for improved layout of the summary view. Changed wording to English in the frontend and updated the routes accordingly.