FMZ-ENGLISH-TRAINER-004 Refactor code for better performance and add statistics feature
In this commit, unused imports have been removed to optimize the performance of the code. The mapping and service classes have been updated to include fields that captures the number of mistakes, correct answers and repetitions made by user.

Also updated the UI components in the Summary view to better represent irregular verb learning progress and performance including average grades and statistics of learned and not learned verbs.

A feature to calculate average grade has been introduced in the Service class. This gives the user a better understanding of their performance over time.

Finally, minor tweaks in the ExamView class to enhance user's experience and to ensure correctness of the learning process.