FMZ-05 Update UI and logic for English Trainer application
Updated frontend CSS for better readability and UI/UX in the exam-style.css file.
Added more CSS classes for buttons and overhauled the look.
Introduced new buttons : "Start Exam", "Submit Exam" and "Summary Board".
This change improves the usability of the app by offering a clear and concise user interface and provides a better visual experience to users.

Renamed `TranslationIrregularVerbService` to the more descriptive `TranslationIrregularVerbSummaryService` and moved some of its logic to `TranslationIrregularVerbExamService` to better encapsulate responsibilities.

Introduced `TranslationIrregularVerbExamService` to handle logic related to the examination process. This includes preparing the exam, checking the answers, calculating the grade and saving the exam summary.

Refactored `TranslationIrregularVerbExamView` and `TranslationIrregularVerbSummaryView` to streamline the creation of layouts and grids, thus reducing repetition and improving code readability.

Overall, these changes are aimed at separating concerns among classes and increasing the readability and maintainability of the code.